refactor(ai-builder): Extract trackWebFetchEvents helper and clean up web-fetch modules
Move web-fetch telemetry tracking into a dedicated trackWebFetchEvents function
in builder.store.ts. Includes file reorganization and cleanup across web-fetch
security, allowed-domains, and discovery types.
