+++ packages/@n8n/ai-workflow-builder.ee/src/types/discovery-types.ts
@@ -30,7 +30,7 @@ export interface DiscoveryContext {
 		availableResources?: ResourceInfo[];
 	}>;
 	bestPractices?: string;
-	/** Content fetched via web_fetch during discovery, keyed by URL */
+	/** Content fetched via web_fetch during discovery */
 	fetchedUrlContent?: Array<{
 		url: string;
 		status: 'success' | 'error';
